fix: use ConfidentialityPolicy enum in JobQuoteDetails instead of raw uint8 (M8)
- Types.sol: change confidentiality field to ConfidentialityPolicy enum
- SignatureLib.sol: cast to uint8 for abi.encode (EIP-712 wire format)
- All tests updated to use Types.ConfidentialityPolicy.Any

+++ src/libraries/SignatureLib.sol
@@ -184,7 +184,7 @@ library SignatureLib {
                 details.price,
                 details.timestamp,
                 details.expiry,
-                details.confidentiality
+                uint8(details.confidentiality)
             )
         );
     }
